+++ docs/releases.rst
@@ -4,6 +4,8 @@ Releases
 Unreleased
 ----------
 **Removed**
+* ``SAInstanceView`` class;
+* ``SAItemView`` synonym for ``SAInstanceView``;
 * ``SAAbstractView`` synonym for ``SAMixin``;
 * ``SAOneModelMixin`` synonym for ``SAModelMixin``;
 * ``SAInstanceAddMixin`` synonym for ``SAItemAddMixin``;